pyproject.toml
[tool.poetry]
name = "ginconfig"
version = "0.1.0"
description = ""
authors = ["Maria Bader <X@X.com>"]
readme = "README.md"
packages = [{include = "ginconfig"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.2"
gin-config = "^0.5.0"
numpy = "^1.25.0"
pydantic = "^2.0"
typer = {extras = ["all"], version = "^0.9.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gin = 'main:cli'



